requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anthropic==0.39.0
APScheduler==3.10.4
anyio==4.4.0
blinker==1.8.2
boto3==1.35.63
botocore==1.35.63
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
fal_client==0.5.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
groq==0.12.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
limits==3.13.0
Mako==1.3.10
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.38.0
ordered-set==4.1.0
packaging==24.1
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.3
secure-smtplib==0.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tinycss2==1.4.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.2.1
